feat: Add document class with file I/O capabilities
- Add document class that owns buffer to ensure string_view validity
- Implement file I/O methods (from_file, to_file) with error handling
- Add file_operation_error exception for I/O operations
- Update documentation (README, INDEX, new DOCUMENT.md)
- Update CHANGELOG with new features

This change extends the API without breaking existing code.

+++ CppOnGoogleTest/cppon_udl_test.cpp
@@ -14,7 +14,7 @@ TEST(UDL, JsonQuickAndFull) {
     auto& nq = q["/n"];
     auto& nf = f["/n"];
     EXPECT_NO_THROW((void)get_cast<int>(nq));
-    EXPECT_NO_THROW((void)get_strict<int>(f["/n"]));
+    EXPECT_NO_THROW((void)get_strict<int64_t>(f["/n"]));
     EXPECT_EQ(std::get<string_view_t>(q["/s"]), "x");
     EXPECT_EQ(std::get<string_view_t>(f["/s"]), "x");
 }
